fix: change Price type from int to float in EventPrice model

+++ app/model/event_price.py
@@ -2,4 +2,4 @@ from pydantic import BaseModel
 
 class EventPrice(BaseModel):
     TicketType: str
-    Price: int
+    Price: float
